Move `FeedImageLoaderSpy` to a shared scope to avoid duplication
