+++ EssentialApp/EssentialAppTests/Helpers/FeedImageLoaderSpy.swift
@@ -1,0 +1,44 @@
+//
+//  FeedImageLoaderSpy.swift
+//  EssentialAppTests
+//
+//  Created by Jair Moreno Gaspar on 11/02/21.
+//
+
+import XCTest
+import EssentialFeed
+
+class FeedImageLoaderSpy: FeedImageDataLoader {
+    
+    private var messages: [(url: URL, completion: (FeedImageDataLoader.Result) -> Void)] = []
+    
+    private(set) var cancelledURLs = [URL]()
+
+    var loadedURLs: [URL] {
+        return messages.map { $0.url }
+    }
+    
+    private struct Task: FeedImageDataLoaderTask {
+        let callback: () -> Void
+        
+        func cancel() {
+            callback()
+        }
+    }
+    
+    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        messages.append((url, completion))
+        return Task { [weak self] in
+            self?.cancelledURLs.append(url)
+        }
+    }
+    
+    func complete(with data: Data, at index: Int = 0) {
+        messages[index].completion(.success(data))
+    }
+    
+    func complete(with error: Error, at index: Int = 0) {
+        messages[index].completion(.failure(error))
+    }
+    
+}
